+++ chain/jsonrpc/src/lib.rs
@@ -34,12 +34,12 @@ use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::AccountId;
 use near_primitives::views::FinalExecutionOutcomeViewEnum;
 
+mod api;
 mod errors;
 mod metrics;
-mod parser;
 
+use api::RpcRequest;
 pub use errors::{RpcFrom, RpcInto};
-use parser::RpcRequest;
 
 #[derive(Serialize, Deserialize, Clone, Copy, Debug)]
 pub struct RpcPollingConfig {
